etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Bachelor Projekt Naumann <BP2023FN1@hpi.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
rdflib = "^7.0.0"
pandas = "^2.1.4"
python-slugify = "^8.0.1"
openai = "^1.7.1"
pylint = "^3.0.3"
pyshacl = "^0.25.0"
pytest = "^8.1.1"
pre-commit = "^3.7.0"
azure-storage-blob = "^12.19.1"
levenshtein = "^0.25.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."poetry-dynamic-versioning"]
include = "**"
